+++ spec/models/concerns/searchable_spec.rb
@@ -26,3 +26,42 @@ RSpec.describe Searchable, :xapian do
     end
   end
 end
+
+RSpec.describe Searchable, 'index lifecycle' do
+  it 'indexes a record when it is created' do
+    user = FactoryBot.create(:user)
+    expect(user.search_documents.count).to eq(1)
+  end
+
+  it 'refreshes indexed content when a record is updated' do
+    user = FactoryBot.create(:user, name: 'Original Name')
+    user.update!(name: 'Updated Name')
+
+    content = user.search_documents.reload.first.raw_admin_content
+    expect(content).to include('Updated Name')
+    expect(content).not_to include('Original Name')
+  end
+
+  it 'removes search documents when the record is destroyed' do
+    user = FactoryBot.create(:user)
+    expect { user.destroy! }.to change(SearchDocument, :count).by(-1)
+  end
+
+  it 'does not index models that are not registered as searchable' do
+    body = FactoryBot.create(:public_body)
+    expect(body.search_documents.count).to eq(0)
+  end
+
+  describe '.reindex_all' do
+    it 'indexes every indexable record' do
+      FactoryBot.create_list(:user, 2)
+      SearchDocument.delete_all
+
+      User.reindex_all
+
+      expect(
+        SearchDocument.where(searchable_type: 'User').count
+      ).to eq(User.count)
+    end
+  end
+end